--- src/main/java/me/fengming/openjs/registry/IRegistration.java
@@ -1,4 +1,0 @@
-package me.fengming.openjs.registry;
-
-public interface IRegistration {
-}
