+++ src/main/java/me/fengming/openjs/registry/OpenJSRegistries.java
@@ -1,5 +1,6 @@
 package me.fengming.openjs.registry;
 
+import me.fengming.openjs.registry.api.TypedRegistries;
 import me.fengming.openjs.script.Binding;
 import me.fengming.openjs.script.ScriptType;
