+++ src/main/java/me/fengming/openjs/registry/api/IRegistration.java
@@ -1,0 +1,4 @@
+package me.fengming.openjs.registry.api;
+
+public interface IRegistration {
+}
