+++ src/main/java/me/fengming/openjs/registry/api/TypedRegistry.java
@@ -1,4 +1,4 @@
-package me.fengming.openjs.registry;
+package me.fengming.openjs.registry.api;
 
 import me.fengming.openjs.script.ScriptType;
