+++ src/main/java/me/fengming/openjs/script/Binding.java
@@ -1,6 +1,6 @@
 package me.fengming.openjs.script;
 
-import me.fengming.openjs.registry.IRegistration;
+import me.fengming.openjs.registry.api.IRegistration;
 
 // Is it necessary for it to exist?
 public record Binding(Object value) implements IRegistration {
